Move Matlab plots to new module matlab.plots
